go sidecar improvement
