+++ f3-sidecar/.gitignore
@@ -41,3 +41,5 @@ Cargo.lock
 *.pdb
 
 # End of https://www.toptal.com/developers/gitignore/api/go,rust
+
+/f3-db
